fix(loop-db): make registration and queries transactional
Parse database mutations into explicit success/error outcomes, bind queries to a registered keyframe identity, preserve the full u64 registration sequence space, and rank finite similarities deterministically. Stage sparse map/graph insertion and removal around database failures, and distinguish stale descriptor responses from current invariant violations.

Evidence: 13 focused keyframe-database tests; cargo test -p kiko-slam --lib (389 passed); cargo test -p kiko-slam --bins (3 passed); cargo clippy -p kiko-slam --all-targets -- -D warnings; cargo fmt --all -- --check; git diff --check.

+++ crates/kiko-slam/src/diagnostics.rs
@@ -1006,7 +1006,7 @@ mod tests {
                 source_snapshot: crate::map::SlamMap::new().snapshot(),
                 error: std::sync::Arc::new(
                     crate::loop_closure::KeyframeDatabaseError::SequenceExhausted {
-                        next_sequence: usize::MAX,
+                        next_sequence: u64::MAX,
                     },
                 ),
             },
